pyproject.toml
[project]
name = "stylecompass-core"
version = "0.1.0"
description = "Style-based image embedding, clustering, and recommendation system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "einops>=0.8.0",
    "faiss-cpu>=1.9.0",
    "ftfy>=6.3.1",
    "huggingface-hub>=0.26.2",
    "matplotlib>=3.9.2",
    "numpy>=2.1.3",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "pytest>=8.3.3",
    "pyyaml>=6.0.2",
    "regex>=2024.11.6",
    "scikit-learn>=1.5.2",
    "scipy>=1.14.1",
    "tensorboard>=2.18.0",
    "timm>=1.0.11",
    "torch>=2.5.1",
    "torchaudio>=2.5.1",
    "torchvision>=0.20.1",
    "wandb>=0.18.6",
]

[project.scripts]
compass-download = "stylecompass_core.scripts.download_tars:main"
compass-prepare = "stylecompass_core.scripts.prepare_batch:main"
compass-clean = "stylecompass_core.scripts.clean_cache:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stylecompass_core"]

